Make mdn-data a pure build time dependency.
The mdn-data package is not being released to npm on any sort of regular
basis. This change is preparing for the ability to update directly from
the github repo using scripts or submodules.

Also, it's best if all the data is in kept in sync because the mdn-data
changes are not following any sort of semver concepts in how they land
schema or data changes.